fastapi/requirements.txt
fastapi
pydantic
uvicorn
pymysql
mysql-connector-python
sqlalchemy
geoalchemy2
shapely
